Cleanup: move various functions into a new "phyleticPattern" utilities file
