+++ Site/webapp/wdkCustomization/js/questions/GroupsByPhyleticPattern/Form.tsx
@@ -1,24 +1,35 @@
 import React, { useCallback, useMemo, useState } from 'react';
 
 import produce from 'immer';
-import { groupBy, mapValues, orderBy, partition } from 'lodash';
 
 import { CheckboxTree, IconAlt, Loading } from 'wdk-client/Components';
 import { LinksPosition } from 'wdk-client/Components/CheckboxTree/CheckboxTree';
 import { makeClassNameHelper } from 'wdk-client/Utils/ComponentUtils';
-import { foldStructure, mapStructure } from 'wdk-client/Utils/TreeUtils';
 import { ParameterGroup } from 'wdk-client/Utils/WdkModel';
 import { Props, SubmitButton } from 'wdk-client/Views/Question/DefaultQuestionForm';
 
 import { EbrcDefaultQuestionForm } from 'ebrc-client/components/questions/EbrcDefaultQuestionForm';
 
 import { useTaxonUiMetadata } from 'ortho-client/hooks/taxons';
-import { TaxonTree } from 'ortho-client/utils/taxons';
+import {
+  ConstraintState,
+  ConstraintStates,
+  PhyleticExpressionUiTree,
+  cxPhyleticExpression,
+  getNextConstraintState,
+  getNodeChildren,
+  getNodeId,
+  makeInitialConstraintStates,
+  makeInitialExpandedNodes,
+  makePhyleticExpression,
+  makePhyleticExpressionUiTree,
+  updateChildConstraintStates,
+  updateParentConstraintStates
+} from 'ortho-client/utils/phyleticPattern';
 
 import './Form.scss';
 
 const cxDefaultQuestionForm = makeClassNameHelper('wdk-QuestionForm');
-const cxPhyleticExpression = makeClassNameHelper('PhyleticExpression');
 
 const PHYLETIC_EXPRESSION_PARAM_NAME = 'phyletic_expression';
 
@@ -79,22 +90,6 @@ interface PhyleticExpressionParameterProps {
   submitting: Props['state']['submitting'];
   updatePhyleticExpressionParam: (newParamValue: string) => void;
 }
-
-interface PhyleticExpressionUiTree extends TaxonTree {
-  children: PhyleticExpressionUiTree[];
-  parent?: PhyleticExpressionUiTree;
-  speciesCount: number;
-}
-
-type ConstraintStates = Record<string, ConstraintState>;
-
-type HomogeneousConstraintState =
-  | 'free'
-  | 'include-at-least-one'
-  | 'include-all'
-  | 'exclude';
-
-type ConstraintState = HomogeneousConstraintState | 'mixed';
 
 function PhyleticExpressionParameter({
   phyleticExpressionTextField,
@@ -197,78 +192,6 @@ function PhyleticExpressionParameter({
   );
 }
 
-function makePhyleticExpressionUiTree(taxonTree: TaxonTree) {
-  const phyleticExpressionUiTree = mapStructure(
-    (node: TaxonTree, mappedChildren: PhyleticExpressionUiTree[]) => ({
-      ...node,
-      children: orderBy(
-        mappedChildren,
-        child => child.species,
-        'desc'
-      ),
-      speciesCount: node.species
-        ? 1
-        : mappedChildren.reduce(
-            (memo, { speciesCount }) => memo + speciesCount,
-            0
-          )
-    }),
-    (node: TaxonTree) => node.children,
-    taxonTree
-  );
-
-  _addParentRefs(phyleticExpressionUiTree, undefined);
-
-  return phyleticExpressionUiTree;
-
-  function _addParentRefs(node: PhyleticExpressionUiTree, parent: PhyleticExpressionUiTree | undefined) {
-    if (parent != null) {
-      node.parent = parent;
-    }
-
-    node.children.forEach(child => {
-      _addParentRefs(child, node);
-    });
-  }
-}
-
-function makeInitialExpandedNodes(phyleticExpressionUiTree: PhyleticExpressionUiTree) {
-  const initialExpandedNodes = [] as string[];
-
-  _traverse(phyleticExpressionUiTree, 0, 1);
-
-  return initialExpandedNodes;
-
-  function _traverse(node: PhyleticExpressionUiTree, depth: number, maxDepth: number) {
-    if (depth <= maxDepth) {
-      initialExpandedNodes.push(getNodeId(node));
-
-      node.children.forEach(child => {
-        _traverse(child, depth + 1, maxDepth);
-      });
-    }
-  }
-}
-
-function makeInitialConstraintStates(phyleticExpressionUiTree: PhyleticExpressionUiTree) {
-  return foldStructure(
-    (constraintStates: ConstraintStates, node: PhyleticExpressionUiTree) => {
-      constraintStates[node.abbrev] = 'free';
-      return constraintStates;
-    },
-    {} as ConstraintStates,
-    phyleticExpressionUiTree
-  );
-}
-
-function getNodeId(node: PhyleticExpressionUiTree) {
-  return node.abbrev;
-}
-
-function getNodeChildren(node: PhyleticExpressionUiTree) {
-  return node.children;
-}
-
 function makeRenderNode(
   constraintStates: ConstraintStates,
   setConstraintStates: (newConstraintStates: ConstraintStates) => void,
@@ -348,134 +271,4 @@ function ConstraintIcon({
   return (
     <span className={className} onClick={onClick}></span>
   );
-}
-
-function getNextConstraintState(currentState: ConstraintState, isSpecies: boolean): HomogeneousConstraintState {
-  if (currentState === 'mixed') {
-    return 'include-all';
-  }
-
-  const stateOrder = isSpecies
-    ? SPECIES_STATE_ORDER
-    : NON_SPECIES_STATE_ORDER;
-
-  const stateIndex = stateOrder.indexOf(currentState);
-
-  return stateOrder[(stateIndex + 1) % stateOrder.length];
-}
-
-const NON_SPECIES_STATE_ORDER = [ 'free', 'include-all', 'include-at-least-one', 'exclude' ] as const;
-const SPECIES_STATE_ORDER = [ 'free', 'include-all', 'exclude' ] as HomogeneousConstraintState[];
-
-function updateParentConstraintStates(
-  node: PhyleticExpressionUiTree,
-  draftConstraintStates: ConstraintStates,
-  changedState: HomogeneousConstraintState
-): void {
-  const parent = node.parent;
-
-  if (parent != null) {
-    const distinctChildConstraintTypes = new Set(
-      parent.children.map(
-        child => draftConstraintStates[child.abbrev]
-      )
-    );
-
-    if (
-      distinctChildConstraintTypes.size === 1 &&
-      changedState !== 'include-at-least-one'
-    ) {
-      draftConstraintStates[parent.abbrev] = changedState;
-    } else {
-      draftConstraintStates[parent.abbrev] = 'mixed';
-    }
-
-    updateParentConstraintStates(parent, draftConstraintStates, changedState);
-  }
-}
-
-function updateChildConstraintStates(
-  node: PhyleticExpressionUiTree,
-  draftConstraintStates: ConstraintStates,
-  changedState: HomogeneousConstraintState
-): void {
-  node.children.forEach(child => {
-    if (changedState === 'include-at-least-one') {
-      draftConstraintStates[child.abbrev] = 'free';
-    } else {
-      draftConstraintStates[child.abbrev] = changedState;
-    }
-
-    updateChildConstraintStates(child, draftConstraintStates, changedState);
-  });
-}
-
-function makePhyleticExpression(
-  phyleticExpressionUiTree: PhyleticExpressionUiTree,
-  constraintStates: ConstraintStates
-) {
-  const nonSpeciesExpressionTerms = [] as string[];
-  const includedSpeciesWithMixedParents = [] as string[];
-  const excludedSpeciesWithMixedParents = [] as string[];
-
-  _traverse(phyleticExpressionUiTree);
-
-  const nonSpeciesSubexpression = nonSpeciesExpressionTerms.length == 0
-    ? undefined
-    : nonSpeciesExpressionTerms.join(' AND ');
-
-  const includedSpeciesSubexpression = includedSpeciesWithMixedParents.length == 0
-    ? undefined
-    : `${includedSpeciesWithMixedParents.join('+')}=${includedSpeciesWithMixedParents.length}T`;
-
-  const excludedSpeciesSubexpression = excludedSpeciesWithMixedParents.length == 0
-    ? undefined
-    : `${excludedSpeciesWithMixedParents.join('+')}=0T`;
-
-  const subexpressions = [
-    nonSpeciesSubexpression,
-    includedSpeciesSubexpression,
-    excludedSpeciesSubexpression
-  ];
-
-  return (
-    subexpressions
-      .filter(subexpression => subexpression != null)
-      .join(' AND ')
-  );
-
-  function _traverse(node: PhyleticExpressionUiTree) {
-    const nextConstraintType = constraintStates[node.abbrev];
-
-    if (nextConstraintType === 'include-all') {
-      nonSpeciesExpressionTerms.push(`${node.abbrev}=${node.speciesCount}T`);
-    } else if (nextConstraintType === 'include-at-least-one') {
-      nonSpeciesExpressionTerms.push(`${node.abbrev}>=1T`);
-    } else if (nextConstraintType === 'exclude') {
-      nonSpeciesExpressionTerms.push(`${node.abbrev}=0T`);
-    } else if (nextConstraintType === 'mixed') {
-      const [ speciesChildren, nonSpeciesChildren ] = partition(
-        node.children,
-        child => child.species
-      );
-
-      const speciesChildrenAbbrevs = mapValues(
-        speciesChildren,
-        speciesChild => speciesChild.abbrev
-      );
-
-      const speciesChildrenAbbrevsByState = groupBy(
-        speciesChildrenAbbrevs,
-        speciesChildAbbrev => constraintStates[speciesChildAbbrev]
-      );
-
-      const includedSpeciesAbbrevs = speciesChildrenAbbrevsByState['include-all'] ?? [];
-      const excludedSpeciesAbbrevs = speciesChildrenAbbrevsByState['exclude'] ?? [];
-
-      includedSpeciesWithMixedParents.push(...includedSpeciesAbbrevs);
-      excludedSpeciesWithMixedParents.push(...excludedSpeciesAbbrevs);
-
-      nonSpeciesChildren.forEach(_traverse);
-    }
-  }
 }
